requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.22
pydantic==2.12.5
websockets==15.0.1
python-dotenv==1.2.1
livekit-api>=0.7.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
google-generativeai>=0.8.0
anyio>=4.12.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
fsspec==2026.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore>=1.0.0
httpx>=0.26.0
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.0
postgrest==2.28.0
propcache==0.4.1
pycparser==3.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
realtime==2.28.0
requests==2.32.5
rich==14.3.3
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vastai-sdk>=0.5.0
yarl==1.22.0
zstandard==0.25.0
ffmpeg-python==0.2.0
av>=12.0.0
Pillow>=10.0.0
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.30.0
